app/requirements.txt
torch
numpy
matplotlib
pillow
gdal
psycopg2-binary
scikit-learn
